requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.2.1
openai==1.34.0
faiss-cpu==1.7.4
tqdm==4.66.4